LPS-91769 Move all client side class into one package
