+++ modules/test/arquillian-extension-junit-bridge/src/main/java/com/liferay/arquillian/extension/junit/bridge/client/JMXProxyUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.arquillian.extension.junit.bridge.jmx;
+package com.liferay.arquillian.extension.junit.bridge.client;
 
 import java.io.IOException;
